refactor(cli): split commands god-module — extract oracle subcommands
commands.rs was the repo's hottest god-module (1673 lines, 50 fns, 31
recent commits). Convert it to a directory module and lift the largest,
most self-contained cluster into commands/oracle.rs:

- commands/oracle.rs: oracle / with_oracle_write_lock / oracle_run /
  oracle_status / oracle_report / ensure_checkout_matches_corpus +
  the provenance/tool-version helpers, plus the oracle tests
  (temp_config, the write-lock discipline tests, the corpus-bindings
  check). All three of the file's repo memories were oracle invariants.
- commands/mod.rs: the remaining handlers, curating the surface via
  `pub(crate) use oracle::{oracle, with_oracle_write_lock}` so main.rs's
  `use commands::*` dispatch (and the background auto-run path) are
  unchanged — no call sites touched.

Pure code movement; behaviour identical. cargo +nightly fmt, clippy
--all-targets, and `cargo nextest run -p rag-rat` (104 tests) green.
